// src/app/models/ariziModels/ariziYolcu.ts
export interface AriziYolcu {
  Id: number;
  UetdsSeferReferansNo: number;
  UetdsGrupRefNo: string;
  UetdsYolcuRefNo: string;
  UyrukUlke: string;
  Cinsiyet: string;
  TcKimlikPasaportNo: string;
  Adi: string;
  Soyadi: string;
  Koltuk: string;
  Telefon: string;
  HesKodu: string;
  AktifIptal: string;
  IptalAciklama: string;
}